scripts/requirements.txt
insightface
onnxruntime
numpy
scikit-learn
opencv-python-headless
ultralytics
whisperx
python-dotenv
transformers
torch
Pillow
